requirements.txt
pytesseract
Pillow
pandas
openpyxl
pdf2image
